style: satisfy rustfmt and clippy across the workspace
First CI run on this branch surfaced accumulated lint drift (the branch only
runs CI via PR). Apply `cargo fmt --all` and resolve clippy lints:

- factor complex types into aliases (browser tabs handler, webview controller
  setup, windows native video params)
- collapse nested ifs into let-chains, use `clamp`, drop needless casts/borrows
  and a useless conversion across the Windows platform/SDK code
- gate `WebViewSession*::is_destroyed` (Apple-only caller) as dead_code-allowed
  off-Apple; move `lingxia-shell` downloads `register()` above its test module
- fix a latent bug clippy caught: `SystemDrive` join produced `\` instead of
  the drive root

+++ crates/lingxia-platform/src/windows/mouse.rs
@@ -153,7 +153,7 @@ fn post_wheel_message(
         let _ = ClientToScreen(target, &mut screen);
     }
     let delta = (amount * f64::from(WHEEL_DELTA)) as i16;
-    let wparam = WPARAM(((delta as u16 as usize) << 16) as usize);
+    let wparam = WPARAM((delta as u16 as usize) << 16);
     let lparam = LPARAM(((screen.y as isize) << 16) | (screen.x as isize & 0xFFFF));
     unsafe {
         WindowsAndMessaging::PostMessageW(Some(target), message, wparam, lparam)
